1- extracted web client, tests with wiremock
