+++ passport-processing/src/main/java/ru/veselov/passportprocessing/service/PdfHttpClient.java
@@ -1,0 +1,9 @@
+package ru.veselov.passportprocessing.service;
+
+import org.springframework.core.io.buffer.DataBuffer;
+
+public interface PdfHttpClient {
+
+    DataBuffer sendRequestForConvertingDocxToPdf(byte[] source);
+
+}
