+++ passport-processing/src/main/java/ru/veselov/passportprocessing/service/impl/PdfServiceImpl.java
@@ -2,74 +2,26 @@ package ru.veselov.passportprocessing.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.buffer.DataBuffer;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.MediaType;
-import org.springframework.http.client.MultipartBodyBuilder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.BodyInserters;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.WebClientRequestException;
-import reactor.core.publisher.Mono;
 import ru.veselov.passportprocessing.exception.PdfProcessingException;
-import ru.veselov.passportprocessing.exception.ServiceUnavailableException;
+import ru.veselov.passportprocessing.service.PdfHttpClient;
 import ru.veselov.passportprocessing.service.PdfService;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.concurrent.atomic.AtomicReference;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class PdfServiceImpl implements PdfService {
 
-    public static final String CONTENT_DISPOSITION = "Content-Disposition";
-
-    @Value("${pdf-service.url}")
-    private String pdfConverterUrl;
-
-    @Value("${pdf-service.filename}")
-    private String filename;
-
-    private final WebClient webClient;
+    private final PdfHttpClient pdfHttpClient;
 
     @Override
     public byte[] createPdf(byte[] source) {
-        MultipartBodyBuilder bodyBuilder = new MultipartBodyBuilder();
-        bodyBuilder.part("file", source)
-                .header(CONTENT_DISPOSITION, "form-data; name=" + filename).filename(filename + ".docx");
-        Mono<DataBuffer> dataBufferMono = webClient.post().uri(pdfConverterUrl)
-                .contentType(MediaType.MULTIPART_FORM_DATA)
-                .body(BodyInserters.fromMultipartData(bodyBuilder.build()))
-                .retrieve()
-                .onStatus(HttpStatusCode::is4xxClientError, response -> {
-                    log.error(
-                            "Can't convert to Pdf, something wrong with generated docx file, [status: {}, message: {}]",
-                            response.statusCode().value(), response.bodyToMono(String.class).block());
-                    throw new PdfProcessingException(
-                            "Can't convert to Pdf, something wrong with generated docx file, status: %s, message: %s"
-                                    .formatted(response.statusCode().value(),
-                                            response.bodyToMono(String.class).block()));
-                })
-                .onStatus(HttpStatusCode::is5xxServerError, response -> {
-                    log.error("Something wrong with Pdf service, [status: {}, message: {}]",
-                            response.statusCode().value(), response.bodyToMono(String.class).block());
-                    throw new ServiceUnavailableException(
-                            "Something wrong with Pdf service, [status: %s, message: %s]".formatted(
-                                    response.statusCode().value(), response.bodyToMono(String.class).block()
-                            )
-                    );
-                })
-                .bodyToMono(DataBuffer.class).doOnError(t -> log.error("Pdf server is down: [{}]", t.getMessage()))
-                .onErrorResume(WebClientRequestException.class::isInstance, e -> {
-                    throw new ServiceUnavailableException("Pdf server is down: %s".formatted(e.getMessage()));
-                });
-
-
-        DataBuffer pdfDatabuffer = dataBufferMono.block();
+        DataBuffer pdfDatabuffer = pdfHttpClient.sendRequestForConvertingDocxToPdf(source);
         log.info("Document successfully converted to pdf");
         return convertToByteArray(pdfDatabuffer);
     }
