+++ passport-processing/src/test/java/ru/veselov/passportprocessing/service/impl/PdfServiceImplTest.java
@@ -1,54 +1,36 @@
 package ru.veselov.passportprocessing.service.impl;
 
-import com.github.tomakehurst.wiremock.WireMockServer;
-import com.github.tomakehurst.wiremock.client.WireMock;
-import lombok.SneakyThrows;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.springframework.test.util.ReflectionTestUtils;
-import org.springframework.web.reactive.function.client.WebClient;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentMatchers;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.core.io.buffer.DataBuffer;
+import ru.veselov.passportprocessing.service.PdfHttpClient;
 
+import java.io.ByteArrayInputStream;
+
+@ExtendWith(MockitoExtension.class)
 class PdfServiceImplTest {
 
-    public static WireMockServer wireMockServer;
+    @Mock
+    PdfHttpClient pdfHttpClient;
 
-    public static int PORT = 30001;
-
-    public PdfServiceImpl pdfService;
-
-    private String pdfConverterUrl;
-
-    @BeforeAll
-    @SneakyThrows
-    static void setUp() {
-        wireMockServer = new WireMockServer(PORT);
-        wireMockServer.start();
-    }
-
-    @AfterAll
-    @SneakyThrows
-    static void shutdown() {
-        wireMockServer.shutdown();
-    }
-
-    @BeforeEach
-    void init() {
-        WebClient webClient = WebClient.create();
-        pdfConverterUrl = "http://localhost:%d".formatted(PORT);
-        pdfService = new PdfServiceImpl(webClient);
-        ReflectionTestUtils.setField(pdfService, "pdfConverterUrl", pdfConverterUrl, String.class);
-    }
+    @InjectMocks
+    PdfServiceImpl pdfService;
 
     @Test
-    @SneakyThrows
-    void shouldCallPdfService() {
-        WireMock.configureFor("localhost", PORT);
-        WireMock.stubFor(WireMock.post("/")
-                .willReturn(WireMock.aResponse().withStatus(200).withBody(new byte[]{1, 2, 3, 4})));
-        pdfService.createPdf(new byte[]{1, 2, 3, 4, 5});
-        WireMock.verify(WireMock.postRequestedFor(WireMock.urlEqualTo("/")));
+    void shouldReturnByteArray() {
+        DataBuffer mockDataBuffer = Mockito.mock(DataBuffer.class);
+        Mockito.when(mockDataBuffer.asInputStream()).thenReturn(new ByteArrayInputStream(new byte[]{1, 2, 3, 4}));
+        Mockito.when(pdfHttpClient.sendRequestForConvertingDocxToPdf(ArgumentMatchers.any()))
+                .thenReturn(mockDataBuffer);
+        byte[] pdfBytes = pdfService.createPdf(new byte[]{1, 2});
+
+        Assertions.assertThat(pdfBytes).isNotNull();
     }
 
 }
